wip - table destructuring
